Use Guzzle/Service/Client get Method to Set Header
Remove calls to `setDefaultOption` to set the headers for the
get request being generated by the AbstractProvider. This change
was made to fix `Requires authentication` errors being recieved
from GitHub and `Unknown authentication scheme` errors being
recieved from LinkedIn.

Looking into the documentation on Guzzle the get method on the
client will manage and set our headers for us, and was used in
place of `setDefaultOption`.

Tests were updated to reflect the removal of the call to
`setDefaultOption`.