Convert all uses of "User" to "Resource Owner"
This is an exploratory change to get feedback from the oauth2-client
community. It is the result of a conversation on Twitter with Neil
Crookes: https://twitter.com/ramsey/status/618437720554143744

I have never been comfortable providing any entity that represents a
user, since the OAuth 2.0 specification does not specify anything to do
with users. However, the specification does talk about resource owners,
which may be end-users, if the resource owner is a person (see section
1.1 of the OAuth 2.0 spec), but it may not be a user at all, as is the
case with Neil's use-case.

With this change, the concept of the "user" is migrated to that of a
"resource owner." Individual providers may call their class based on the
new ResourceOwnerInterface a User, if the resource owner in their
service is a person, or they may call it something else entirely.