Storing the Provider Response in a class attribute
Would like to propose storing the response in a class attribute, so
that if we configure the OAuth to get additional information than the
User Object has, then this attribute can be used to get that
information. For instance I was trying to get “summary” and “Compaines”
fields from LinkedIn, I am able to set $fields to include that in the
call to linked in but the return object only returns the User object
and that does not have all the fields. So providing a way to access the
actual response received would be good so that users can use this
extended information