getRandomState only returns alphanumeric states
Before, this was using the default character set for
RandomLib/Generator's generateString which is the base 64 character set
that includes + and /. While / wasn't causing any problems, using + in a
URL parameters (e.g. when the OAuth 2 server sends back the state), the
+ was getting interpretted as a space, which means when a straight
string comparison to stored state was being done, it was returning
false.

This changes getRandomState to use the Generator::CHAR_ALNUM constant as
its character set which solves this problem.