Allow fetching unparsed responses
OAuth2 Client providers generally make good API gateways, due to the
inclusion of `getAuthenticatedRequest`. However, the `getResponse`
method always assumes that:

1. The response can be parsed by the provider.
2. The response headers can be discarded.

These two assumptions make it hard to use as a general purpose API
gateway in many situations where content does not come back as JSON or
when headers are required.

Thanks to @rtheunissen for most of this code, which was adapted for
changes made in version 2.x.

Fixes #408